src/tensorrt/tensorrt-6.0.1.5/samples/python/uff_ssd/requirements.txt
numpy
Pillow
pycuda
requests
tensorflow
